Handle it for renames too
Took 2 minutes

+++ src/main/java/me/glaremasters/guilds/commands/CommandGuilds.java
@@ -77,16 +77,7 @@ public class CommandGuilds extends BaseCommand {
             }
         }
 
-        if (guilds.getConfig().getBoolean("enabled-blacklist")) {
-            List<String> blacklist = guilds.getConfig().getStringList("blacklist");
-
-            for (String censor : blacklist) {
-                if (name.toLowerCase().contains(censor)) {
-                    getCurrentCommandIssuer().sendInfo(Messages.ERROR__BLACKLIST);
-                    return;
-                }
-            }
-        }
+        if (check(name)) return;
         getCurrentCommandIssuer().sendInfo(Messages.CREATE__WARNING);
 
         guilds.getActionHandler().addAction(player, new ConfirmAction() {
@@ -218,6 +209,23 @@ public class CommandGuilds extends BaseCommand {
             getCurrentCommandIssuer().sendInfo(Messages.ERROR__ROLE_NO_PERMISSION);
             return;
         }
+        int minLength = guilds.getConfig().getInt("name.min-length");
+        int maxLength = guilds.getConfig().getInt("name.max-length");
+        String regex = guilds.getConfig().getString("name.regex");
+
+        if (name.length() < minLength || name.length() > maxLength || !name.matches(regex)) {
+            getCurrentCommandIssuer().sendInfo(Messages.CREATE__REQUIREMENTS);
+            return;
+        }
+
+        for (String guildName : guilds.getGuildHandler().getGuilds().keySet()) {
+            if (guildName.equalsIgnoreCase(name)) {
+                getCurrentCommandIssuer().sendInfo(Messages.CREATE__GUILD_NAME_TAKEN);
+                return;
+            }
+        }
+
+        if (check(name)) return;
         String oldName = guild.getName();
         guilds.getDatabase().removeGuild(Guild.getGuild(oldName));
         getCurrentCommandIssuer().sendInfo(Messages.RENAME__SUCCESSFUL, "{name}", name);
@@ -822,6 +830,20 @@ public class CommandGuilds extends BaseCommand {
 
         skull.setItemMeta(meta);
         return skull;
+    }
+
+    private boolean check(String name) {
+        if (guilds.getConfig().getBoolean("enabled-blacklist")) {
+            List<String> blacklist = guilds.getConfig().getStringList("blacklist");
+
+            for (String censor : blacklist) {
+                if (name.toLowerCase().contains(censor)) {
+                    getCurrentCommandIssuer().sendInfo(Messages.ERROR__BLACKLIST);
+                    return true;
+                }
+            }
+        }
+        return false;
     }
 
 }
